+++ src/JavLuv/MovieDetailViewModel.cs
@@ -388,8 +388,9 @@ namespace JavLuv
                 {
                     m_movieData.Metadata.Actors = actors;
                     m_movieData.MetadataChanged = true;
-                    Actresses = new MovieDetailActressViewModel(this);
                     NotifyPropertyChanged("Actors");
+                    Actresses = new MovieDetailActressViewModel(this);
+                    NotifyPropertyChanged("Actresses");
                 }
             }
         }